standup/requirements.txt
asgiref==3.7.2
black==23.9.0
build==1.0.3
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.7
crispy-bootstrap5==0.7
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-htmx==1.16.0
django-password-reset==2.0
django-postgres-extensions==0.9.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.12.3
gunicorn==21.2.0
identify==2.5.27
idna==3.4
install==1.3.5
logtail-python==0.2.6
loguru==0.7.1
msgpack==1.0.6
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
pendulum==2.1.2
platformdirs==3.10.0
pre-commit==3.4.0
psycopg2-binary==2.9.7
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.5
virtualenv==20.24.5
whitenoise==6.5.0
